noise is part of drawing?
