+++ bpd/draw.py
@@ -1,5 +1,9 @@
 import galsim
+import jax.numpy as jnp
 import jax_galsim as xgalsim
+from jax import random
+from jax._src.prng import PRNGKeyArray
+from jax.typing import ArrayLike
 from jax_galsim import GSParams
 
 
@@ -50,3 +54,20 @@ def draw_gaussian_galsim(
     gal_conv = galsim.Convolve([gal, psf])
     image = gal_conv.drawImage(nx=slen, ny=slen, scale=pixel_scale, offset=(x, y))
     return image.array
+
+
+def add_noise(
+    rng_key: PRNGKeyArray,
+    x: ArrayLike,
+    bg: float,
+    n: int = 1,
+):
+    """Produce `n` independent Gaussian noise realizations of a given image `x`.
+
+    NOTE: This function assumes image is background-subtracted and dominated.
+    """
+    assert isinstance(bg, float) or bg.shape == ()
+    x = x.reshape(1, *x.shape)
+    x = x.repeat(n, axis=0)
+    noise = random.normal(rng_key, shape=x.shape) * jnp.sqrt(bg)
+    return x + noise
